requirements.txt
PyMuPDF==1.22.5
pillow==10.0.0
requests==2.31.0